add testing

+++ app.py
@@ -346,8 +346,23 @@ def messages_add():
 def messages_show(message_id):
     """Show a message."""
 
-    msg = Message.query.get(message_id)
+    msg = Message.query.get_or_404(message_id)
     return render_template('messages/show.html', message=msg)
+
+
+# @app.route('/messages/<int:message_id>/delete', methods=["POST"])
+# def messages_destroy(message_id):
+#     """Delete a message."""
+
+#     if not g.user:
+#         flash("Access unauthorized.", "danger")
+#         return redirect("/")
+
+#     msg = Message.query.get(message_id)
+#     db.session.delete(msg)
+#     db.session.commit()
+
+#     return redirect(f"/users/{g.user.id}")
 
 
 @app.route('/messages/<int:message_id>/delete', methods=["POST"])
@@ -359,6 +374,11 @@ def messages_destroy(message_id):
         return redirect("/")
 
     msg = Message.query.get(message_id)
+
+    if msg.user.id != g.user.id:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
+
     db.session.delete(msg)
     db.session.commit()
